Started formalizing some common type classes
One of the first things that come to mind when discussing the
shortcomings of the Coq standard library is its lack of type classes
that are present in Haskell's Prelude and make the developers' lives
much easier by allowing them to enhance a data structure with a rich set
of common functions just by describing its categorical properties.

The algorithms that shall be used in this work aren't trivial. There
will be a substantial amount of parsing, for example, which is a painful
process when performed without help of some Alternative and Applicative
instances.

This commit introduces Functor, Applicative, and Alternative type
classes as well as their laws and an `option` instance for them.